requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
bleach==5.0.1
cssselect==1.2.0
Django==3.2
django-markdownfield==0.10.0
django-social-share==2.3.0
django-taggit==3.1.0
djangorestframework==3.14.0
lxml==4.9.2
Markdown==3.4.1
markdown2==2.4.6
Pillow==9.3.0
pyquery==1.4.3
pytz==2022.7
readtime==2.0.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
timeago==1.0.16
webencodings==0.5.1
